Refactored modules in anticipation of adding Inventory rooted urls
Today we mostly learned about childInjectors.

They weren't what I thought, as there is no notion of 'abstract'. Each injector must be complete. This has meant that only modules without dependencies can be in the parent. In this instance that only really leaves the persistence layer... which is actually fine by me.